hepsiburadacontentgen-feature-scrape-and-image-2/requirements.txt
flask==2.3.3
requests==2.31.0
openai==1.3.0
python-dotenv==1.0.0
urllib3==1.26.15
Pillow==9.5.0
gunicorn==21.2.0
beautifulsoup4==4.12.2
supabase==1.0.3
fal-client==0.5.9
replicate==0.22.0
boto3==1.28.38
# Replacing problematic packages
# scrapeapi-client may be causing issues 